Fix the "git statu" bug (sed portability)
[The `\s` is a GNU sed extension][sed] and is therefore not portable.
On some systems (including OS X) `\s` is interpreted as a character `s`
and therefore commands that end with `s` (such as `git status` or `git
status -s`) will get the last letter stripped. For example, `gst`
becomes `git statu`.

The issue can easily be avoided with the more portable `[:space:]`
character class that is part of the POSIX standard.

[sed]: http://stackoverflow.com/a/10422020/2079797