Add option to ggpush abbreviation
`ggpush` is `git push origin HEAD` which is "a handy way to push the
current branch to the same name on the remote." (from man git-push) The
`--set-upstream` option adds an upstream (tracking) reference, enabling
future commands like `git push` work without any arguments. This fits
nicely to the spirit of `ggpush`.